api/requirements.txt
django>=3.2
psycopg2>=2.9.1
djangorestframework>=3.12.4
djangorestframework-simplejwt
django-cors-headers
pyyaml